#60 検索機能の追加、リセットボタン
