+++ src/components/TodoList.tsx
@@ -35,57 +35,56 @@ export type Todo = {
 };
 
 type TodoListProps = {
-  filterStatus: FilterStatus;
-  filterPriority: FilterPriority;
+  filteredTodos: Todo[];
 }
-const TodoList: React.FC<TodoListProps> = ({filterStatus, filterPriority}) => {
+const TodoList: React.FC<TodoListProps> = ({filteredTodos}) => {
   const router = useRouter();
-  const [todos, setTodos] = useState<Todo[]>([]);
-  const [filteredTodos, setFilteredTodos] = useState<Todo[]>([])
+  // const [todos, setTodos] = useState<Todo[]>([]);
+  // const [filteredTodos, setFilteredTodos] = useState<Todo[]>([])
 
-  const formatDate = (date: Timestamp) => {
-    const toDatedDate = date.toDate();
-    return dayjs(toDatedDate).format("YYYY-MM-DD HH:mm");
-  }
+  // const formatDate = (date: Timestamp) => {
+  //   const toDatedDate = date.toDate();
+  //   return dayjs(toDatedDate).format("YYYY-MM-DD HH:mm");
+  // }
 
-  useEffect(() => {
-    const q = query(collection(db, "todos"), orderBy("created_at"));
-    onSnapshot(q, (snapshot) => {
-      setTodos(
-        snapshot.docs.map((doc) => {
-          return {
-            id: doc.id,
-            title: doc.data().title as string,
-            detail: doc.data().detail as string,
-            status: doc.data().status as "NOT STARTED" | "DOING" | "DONE",
-            priority: doc.data().priority as "High" | "Middle" | "Low",
-            draft: doc.data().draft as boolean,
-            created_at: formatDate(doc.data().created_at),
-            updated_at: formatDate(doc.data().updated_at)
-          };
-        })
-      );
-    });
-  },[]);
+  // useEffect(() => {
+  //   const q = query(collection(db, "todos"), orderBy("created_at"));
+  //   onSnapshot(q, (snapshot) => {
+  //     setTodos(
+  //       snapshot.docs.map((doc) => {
+  //         return {
+  //           id: doc.id,
+  //           title: doc.data().title as string,
+  //           detail: doc.data().detail as string,
+  //           status: doc.data().status as "NOT STARTED" | "DOING" | "DONE",
+  //           priority: doc.data().priority as "High" | "Middle" | "Low",
+  //           draft: doc.data().draft as boolean,
+  //           created_at: formatDate(doc.data().created_at),
+  //           updated_at: formatDate(doc.data().updated_at)
+  //         };
+  //       })
+  //     );
+  //   });
+  // },[]);
 
-  useEffect(() => {
-    if (filterStatus && filterPriority) {
-      const result = todos
-      .filter((compareStatus) => compareStatus.status === filterStatus)
-      .filter((comparePriority) => comparePriority.priority === filterPriority);
-      setFilteredTodos(result);
-    } else if (!filterStatus && filterPriority) {
-      const result = todos
-      .filter((comparePriority) => comparePriority.priority === filterPriority);
-      setFilteredTodos(result);
-    } else if (filterStatus && !filterPriority) {
-      const result = todos
-      .filter((compareStatus) => compareStatus.status === filterStatus);
-      setFilteredTodos(result);
-    } else {
-      setFilteredTodos(todos);
-    }
-  }, [filterStatus, filterPriority, todos]);
+  // useEffect(() => {
+  //   if (filterStatus && filterPriority) {
+  //     const result = todos
+  //     .filter((compareStatus) => compareStatus.status === filterStatus)
+  //     .filter((comparePriority) => comparePriority.priority === filterPriority);
+  //     setFilteredTodos(result);
+  //   } else if (!filterStatus && filterPriority) {
+  //     const result = todos
+  //     .filter((comparePriority) => comparePriority.priority === filterPriority);
+  //     setFilteredTodos(result);
+  //   } else if (filterStatus && !filterPriority) {
+  //     const result = todos
+  //     .filter((compareStatus) => compareStatus.status === filterStatus);
+  //     setFilteredTodos(result);
+  //   } else {
+  //     setFilteredTodos(todos);
+  //   }
+  // }, [filterStatus, filterPriority, todos]);
 
   const handleEdit = async (selectedId: string) => {
     const docRef = doc(db, "todos", selectedId);
